+++ index.html
@@ -8,16 +8,12 @@
             rel="stylesheet"
         />
         <link
-            href="https://fonts.googleapis.com/css2?family=Kosugi+Maru&family=Patrick+Hand&display=swap"
-            rel="stylesheet"
-        />
-        <link
             href="https://fonts.googleapis.com/css2?family=Sawarabi+Gothic&display=swap"
             rel="stylesheet"
         />
 
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Vite + React</title>
+        <title>NextPlan</title>
     </head>
     <body>
         <div id="root"></div>
