+++ src/components/Login.jsx
@@ -1,0 +1,41 @@
+import React, { useState } from "react";
+import { auth } from "../firebaseConfig";
+import { signInWithEmailAndPassword } from "firebase/auth";
+
+const Login = () => {
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+
+    const handleLogin = async (e) => {
+        e.preventDefault();
+        try {
+            await signInWithEmailAndPassword(auth, email, password);
+            // ログイン成功
+        } catch (error) {
+            console.error("ログインエラー:", error);
+        }
+    };
+
+    return (
+        <form onSubmit={handleLogin}>
+            <h2>ログイン</h2>
+            <input
+                type="email"
+                placeholder="メールアドレス"
+                value={email}
+                onChange={(e) => setEmail(e.target.value)}
+                required
+            />
+            <input
+                type="password"
+                placeholder="パスワード"
+                value={password}
+                onChange={(e) => setPassword(e.target.value)}
+                required
+            />
+            <button type="submit">ログイン</button>
+        </form>
+    );
+};
+
+export default Login;
