+++ src/components/Signup.jsx
@@ -1,0 +1,41 @@
+import React, { useState } from "react";
+import { auth } from "../firebaseConfig";
+import { createUserWithEmailAndPassword } from "firebase/auth";
+
+const Signup = () => {
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+
+    const handleSignup = async (e) => {
+        e.preventDefault();
+        try {
+            await createUserWithEmailAndPassword(auth, email, password);
+            // サインアップ成功
+        } catch (error) {
+            console.error("サインアップエラー:", error);
+        }
+    };
+
+    return (
+        <form onSubmit={handleSignup}>
+            <h2>サインアップ</h2>
+            <input
+                type="email"
+                placeholder="メールアドレス"
+                value={email}
+                onChange={(e) => setEmail(e.target.value)}
+                required
+            />
+            <input
+                type="password"
+                placeholder="パスワード"
+                value={password}
+                onChange={(e) => setPassword(e.target.value)}
+                required
+            />
+            <button type="submit">登録</button>
+        </form>
+    );
+};
+
+export default Signup;
