+++ src/index.css
@@ -98,7 +98,7 @@ body, h1, h2, h3, h4, h5, h6, p, span, label, button {
 
 .dashboard-button:hover {
   background-color: #f9e4b7; /* ホバー時に少し濃い色 */
-  color: #ffffff;
+  color: #777777;
 }
 
 /* タイムライン */
